+++ server/src/main/java/org/allaymc/server/entity/type/EntityTypeInitializer.java
@@ -68,7 +68,7 @@ public final class EntityTypeInitializer {
                         return 0.04;
                     }
                 }, EntityPhysicsComponentImpl.class)
-                .addComponent(EntityAgeComponentImpl::new, EntityAgeComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
                 .build();
     }
 
@@ -116,7 +116,7 @@ public final class EntityTypeInitializer {
                         return 0.04;
                     }
                 }, EntityPhysicsComponentImpl.class)
-                .addComponent(EntityAgeComponentImpl::new, EntityAgeComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
                 .build();
     }
 
@@ -136,7 +136,7 @@ public final class EntityTypeInitializer {
                 .addComponent(EntityProjectileBaseComponentImpl::new, EntityProjectileBaseComponentImpl.class)
                 .addComponent(EntitySnowballPhysicsComponentImpl::new, EntitySnowballPhysicsComponentImpl.class)
                 .addComponent(EntityProjectileComponentImpl::new, EntityProjectileComponentImpl.class)
-                .addComponent(EntityAgeComponentImpl::new, EntityAgeComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
                 .build();
         EntityTypes.SPLASH_POTION = AllayEntityType
                 .builder(EntitySplashPotionImpl.class)
@@ -144,7 +144,7 @@ public final class EntityTypeInitializer {
                 .addComponent(EntityProjectileBaseComponentImpl::new, EntityProjectileBaseComponentImpl.class)
                 .addComponent(EntitySplashPotionPhysicsComponentImpl::new, EntitySplashPotionPhysicsComponentImpl.class)
                 .addComponent(EntitySplashPotionProjectileComponentImpl::new, EntitySplashPotionProjectileComponentImpl.class)
-                .addComponent(EntityAgeComponentImpl::new, EntityAgeComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
                 .build();
         EntityTypes.ENDER_PEARL = AllayEntityType
                 .builder(EntityEnderPearlImpl.class)
@@ -152,7 +152,7 @@ public final class EntityTypeInitializer {
                 .addComponent(EntityProjectileBaseComponentImpl::new, EntityProjectileBaseComponentImpl.class)
                 .addComponent(EntityEnderPearlPhysicsComponentImpl::new, EntityEnderPearlPhysicsComponentImpl.class)
                 .addComponent(EntityProjectileComponentImpl::new, EntityProjectileComponentImpl.class)
-                .addComponent(EntityAgeComponentImpl::new, EntityAgeComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
                 .build();
         EntityTypes.XP_BOTTLE = AllayEntityType
                 .builder(EntityXpBottleImpl.class)
@@ -160,7 +160,7 @@ public final class EntityTypeInitializer {
                 .addComponent(EntityProjectileBaseComponentImpl::new, EntityProjectileBaseComponentImpl.class)
                 .addComponent(EntityXpBottlePhysicsComponentImpl::new, EntityXpBottlePhysicsComponentImpl.class)
                 .addComponent(EntityProjectileComponentImpl::new, EntityProjectileComponentImpl.class)
-                .addComponent(EntityAgeComponentImpl::new, EntityAgeComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
                 .build();
         EntityTypes.ARROW = AllayEntityType
                 .builder(EntityArrowImpl.class)
@@ -168,7 +168,7 @@ public final class EntityTypeInitializer {
                 .addComponent(EntityArrowBaseComponentImpl::new, EntityArrowBaseComponentImpl.class)
                 .addComponent(EntityArrowPhysicsComponentImpl::new, EntityArrowPhysicsComponentImpl.class)
                 .addComponent(EntityProjectileComponentImpl::new, EntityProjectileComponentImpl.class)
-                .addComponent(EntityAgeComponentImpl::new, EntityAgeComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
                 .build();
     }
 }